+++ test/babashka/interop_test.clj
@@ -445,3 +445,31 @@
   (.update standalone-digest data)
 
   (= (seq (.digest standalone-digest)) (seq (.digest sink-digest))))"))))
+
+(deftest java-util-concurrent-locks-test
+  (is (true? (bb nil "(ns script
+  (:import [java.util.concurrent.locks Condition Lock ReentrantLock]
+           [java.util.concurrent TimeUnit]))
+
+(let [result     (volatile! 1)
+      ^Lock lock (ReentrantLock.)
+      cnd        (.newCondition lock)]
+  (try (.lock lock)
+       (future
+         (.lock lock)
+         (vswap! result inc)
+         (.signal cnd)
+         (.unlock lock))
+       (Thread/sleep 1)
+       (.await cnd 10 TimeUnit/MILLISECONDS)
+       (vswap! result * 2)
+       (finally (.unlock lock)))
+  (= 4 @result))")))
+  (is (true? (bb nil "(ns script
+  (:import [java.util.concurrent.locks LockSupport]))
+
+(let [thread (Thread/currentThread)
+      t0     (System/nanoTime)]
+  (future (Thread/sleep 10) (LockSupport/unpark thread))
+  (LockSupport/parkNanos 1e9)
+  (< (- (System/nanoTime) t0) 5e8))"))))